+++ Tests/Device.Tests.ps1
@@ -2,43 +2,8 @@ $credentials = New-Object -TypeName System.Management.Automation.PSCredential -A
 Connect-OMEServer -Name $Global:OMEServer -Credentials $credentials -IgnoreCertificateWarning
 Describe "Device Tests" {
     BeforeEach {
-        $TestDeviceHosts = @("100.79.6.22", "100.79.6.27", "100.79.6.63")
-        $TestDeviceHostsNew = @("100.79.6.64")
         $TestDeviceServiceTags = @("37KP0Q2")
-        $TestDiscoveryJobName = "TestDiscovery01"
         $TestDeviceModel = "PowerEdge R740"
-        $TestiDRACUsername = $Global:iDRACUsername
-        $TestiDRACPassword = $Global:iDRACPassword
-    }
-    Context "Discovery Job Checks" {
-        It "NewDiscoveryJob > Should return a discovery job id" {
-            New-OMEDiscovery -Name $TestDiscoveryJobName -Hosts $TestDeviceHosts -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force) -Wait | Should -BeGreaterThan 0
-        }
-
-        It "GetDiscoveryJob > Should return 1 discovery job" {
-            $TestDiscoveryJobName | Get-OMEDiscovery | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 1
-        }
-
-        It "EditDiscoveryJobAppend > Should return 4 targets hosts" {
-            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Hosts $TestDeviceHostsNew -Mode "Append" -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
-            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 4
-        }
-
-        It "EditDiscoveryJobRemove > Should return 3 target hosts" {
-            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Hosts $TestDeviceHostsNew -Mode "Remove" -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
-            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 3
-        }
-
-        It "EditDiscoveryJobKeepHosts > Should return 3 target hosts" {
-            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
-            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 3
-        }
-
-        It "EditDiscoveryJobRunLater > Should return matching cron string" {
-            $TestCron = "0 0 0 ? * sun *"
-            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Schedule "RunLater" -ScheduleCron $TestCron -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
-            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Schedule | Select-Object -ExpandProperty Cron | Should -Be $TestCron
-        }
     }
     Context "Device Checks" {
         It "Should return ALL Devices" {
@@ -59,6 +24,10 @@ Describe "Device Tests" {
 
         It "Should return device inventory" {
             $TestDeviceServiceTags | Get-OMEDevice | Get-OMEDeviceDetail -InventoryType "deviceSoftware" | Measure-Object | Select-Object -ExpandProperty Count | Should -BeGreaterThan 0
+        }
+
+        It "Should refresh device inventory" {
+            $TestDeviceServiceTags | Get-OMEDevice | Invoke-OMEInventoryRefresh | Should -BeOfType System.Int64
         }
     }
 }